refactor: delete models package
