--- pkg/models/account.go
@@ -1,9 +1,0 @@
-package models
-
-// Account interface that represents a Pocket Network Protocol Account
-type Account interface {
-	GetAddress() string
-	GetBalance() string
-	GetPublicKey() string
-	GetPrivateKey() string
-}
