--- pkg/models/account_with_transactions.go
@@ -1,8 +1,0 @@
-package models
-
-// AccountWithTransactions interface is a wrapper of an account and its transactions
-type AccountWithTransactions interface {
-	Account
-	GetTotalTransactionCount() int
-	GetTransactions() []interface{}
-}
