--- pkg/models/app.go
@@ -1,11 +1,0 @@
-package models
-
-// App interface that represents an app
-type App interface {
-	GetAddress() string
-	GetChains() []string
-	GetMaxRelays() string
-	GetPublicKey() string
-	GetStakedTokens() string
-	GetStatus() StakingStatus
-}
