--- pkg/models/block.go
@@ -1,5 +1,0 @@
-package models
-
-// Block not implemented
-// TODO: implement Block interface
-type Block interface{}
