--- pkg/models/dispatch_request.go
@@ -1,6 +1,0 @@
-package models
-
-// DispatchRequest interface represents a dispatch request
-type DispatchRequest interface {
-	GetSessionHeader() SessionHeader
-}
