--- pkg/models/dispatch_response.go
@@ -1,7 +1,0 @@
-package models
-
-// DispatchResponse interface represents a dispatch response
-type DispatchResponse interface {
-	GetBlockHeight() int
-	GetSession() Session
-}
