--- pkg/models/get_app_options.go
@@ -1,5 +1,0 @@
-package models
-
-// GetAppOptions not implemented
-// TODO: implement GetAppOptions interface
-type GetAppOptions interface{}
