--- pkg/models/get_nodes_options.go
@@ -1,5 +1,0 @@
-package models
-
-// GetNodesOptions not implemented
-// TODO: implement GetNodesOptions interface
-type GetNodesOptions interface{}
