--- pkg/models/node.go
@@ -1,13 +1,0 @@
-package models
-
-// Node interface that represents a node
-type Node interface {
-	GetAddress() string
-	GetChains() []string
-	IsJailed() bool
-	GetPublicKey() string
-	GetServiceURL() string
-	GetStakedTokens() string
-	GetStatus() StakingStatus
-	GetUnstakingTime() string
-}
