--- pkg/models/pocket_aat.go
@@ -1,9 +1,0 @@
-package models
-
-// PocketAAT interface that represents a Pocket Application Authentication Token
-type PocketAAT interface {
-	GetVersion() string
-	GetClientPublicKey() string
-	GetApplicationPublicKey() string
-	GetApplicationSignature() string
-}
