--- pkg/models/relay_headers.go
@@ -1,4 +1,0 @@
-package models
-
-// RelayHeaders map of relay headers
-type RelayHeaders map[string]string
