--- pkg/models/relay_meta.go
@@ -1,6 +1,0 @@
-package models
-
-// RelayMeta interface that represents relay meta
-type RelayMeta interface {
-	GetBlockHeight() string
-}
