--- pkg/models/relay_payload.go
@@ -1,9 +1,0 @@
-package models
-
-// RelayPayload interface that represents
-type RelayPayload interface {
-	GetData() string
-	GetMethod() string
-	GetPath() string
-	GetHeaders() RelayHeaders
-}
