--- pkg/models/relay_proof.go
@@ -1,12 +1,0 @@
-package models
-
-// RelayProof interface represents a relay proof
-type RelayProof interface {
-	GetEntropy() string
-	GetSessionBlockHeight() int
-	GetServicerPubKey() string
-	GetBlockchain() string
-	GetToken() PocketAAT
-	GetSignature() string
-	GetRequestHash() string
-}
