--- pkg/models/relay_request.go
@@ -1,8 +1,0 @@
-package models
-
-// RelayRequest interface represents a relay request
-type RelayRequest interface {
-	GetPayload() RelayPayload
-	GetMeta() RelayMeta
-	GetProof() RelayProof
-}
