--- pkg/models/relay_response.go
@@ -1,9 +1,0 @@
-package models
-
-// RelayResponse interface that represents relay response
-type RelayResponse interface {
-	GetSignature() string
-	GetPayload() string
-	GetProof() RelayProof
-	GetRelayRequest() RelayRequest
-}
