--- pkg/models/request_hash.go
@@ -1,7 +1,0 @@
-package models
-
-// RequestHash interface that represents a request hash
-type RequestHash interface {
-	GetPayload() RelayPayload
-	GetMeta() RelayMeta
-}
