--- pkg/models/session.go
@@ -1,9 +1,0 @@
-package models
-
-// Session interface that represents a session
-type Session interface {
-	GetBlockHeight() int
-	GetHeader() SessionHeader
-	GetKey() string
-	GetNodes() []Node
-}
