--- pkg/models/session_header.go
@@ -1,8 +1,0 @@
-package models
-
-// SessionHeader interface that represent session header
-type SessionHeader interface {
-	GetApplicationPubKey() string
-	GetChain() string
-	GetSessionBlockHeight() int
-}
