--- pkg/models/transaction_response.go
@@ -1,14 +1,0 @@
-package models
-
-// TransactionResponse interface that represents the response of a transaction
-type TransactionResponse interface {
-	GetCode() int
-	GetCodeSpace() string
-	GetData() string
-	GetHash() string
-	GetHeight() int
-	GetInfo() string
-	GetRawLog() string
-	GetTimestamp() string
-	GetTx() string
-}
